refactor: move run_test.py to utils to resolve issue where import from tests occurred
