+++ tests/test_run_tests.py
@@ -4,7 +4,7 @@ import os
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../')))
 
-import src.tests.run_tests as run_tests
+import utils.run_tests as run_tests
 
 
 class TestRunTests(unittest.TestCase):
